test(toolutil): consolidate metatool coverage tests into metatool_test.go
Move 19 tests from metatool_coverage_test.go into metatool_test.go to match the project convention of one test file per source file. No behavior changes; coverage stays at 96.5%.
